Admins can CRUD all notes
