+++ src/components/TaskTable.jsx
@@ -5,10 +5,13 @@ import { Table, Th, Td } from '../styled-components/html';
 import { getNotes } from '../reducers/notesReducer';
 import DeleteNoteButton from './DeleteNoteButton';
 import EditNoteButton from './EditNoteButton';
+import { userStatus } from '../userStatus';
 
 const TaskTable = (props) => {
   const notes = useSelector((state) => state.notes);
   const token = useSelector((state) => state.user.token);
+  const status = useSelector((state) => state.user.status);
+  const isAdmin = status === userStatus.ADMIN;
 
   useEffect(() => {
     (async function getNotesFromServer() {
@@ -20,8 +23,10 @@ const TaskTable = (props) => {
     ? notes.map((note) => (
       <tr key={note.id}>
         <Td>{note.date}</Td>
-        <Td>{note.hours}</Td>
+        {isAdmin
+          && <Td>{note.user ? note.user.username : null}</Td>}
         <Td>{note.task}</Td>
+        <Td>{note.hours}</Td>
         <Td>
           <EditNoteButton note={note} />
           <DeleteNoteButton noteId={note.id} />
@@ -35,8 +40,10 @@ const TaskTable = (props) => {
       <thead>
         <tr>
           <Th>Date</Th>
-          <Th>Hours</Th>
+          {isAdmin
+            && <Th>Username</Th>}
           <Th>Task</Th>
+          <Th>Hours</Th>
           <Th>Actions</Th>
         </tr>
       </thead>
