--- src/styled-components/CancelButton.jsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components';
-import { Button } from './html';
-
-const CancelButton = styled(Button)`
-  background-color: #9e0d0d;
-  margin-left: 12px;
-`;
-
-export default CancelButton;
